Renamed component files
